Refactor JythonWrapperConfiguration and fix label on JGitResourceRepository
