+++ python-app-starters-common/src/main/java/org/springframework/cloud/stream/app/common/resource/repository/JGitResourceRepository.java
@@ -318,6 +318,7 @@ public class JGitResourceRepository implements InitializingBean {
 			if (git != null) {
 				git.close();
 			}
+			refresh();
 		}
 
 	}
