+++ python-app-starters-common/src/main/java/org/springframework/cloud/stream/app/python/wrapper/JythonWrapperProperties.java
@@ -14,11 +14,11 @@
  *   limitations under the License.
  */
 
-package org.springframework.cloud.stream.app.python.local.wrapper;
+package org.springframework.cloud.stream.app.python.wrapper;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.cloud.stream.app.python.jython.JythonScriptProperties;
-import org.springframework.cloud.stream.app.python.script.ScriptProperties;
+
 /**
  * Configuration properties for the Jython wrapper.
  *
